+++ outputs.tf
@@ -7,3 +7,8 @@ output "s3_bucket_name" {
   description = "S3 bucket name"
   value       = aws_s3_bucket.this.bucket
 }
+
+output "cloudfront_distribution_id" {
+  description = "CloudFront distribution ID"
+  value       = aws_cloudfront_distribution.this.id
+}